Updated some extra things for Button in UIToolkit, made a basic menu implementation with some examples how to use Button
