+++ Project 2/Game.py
@@ -2,15 +2,14 @@
 from threading import Thread
 import os, pygame
 import time
-from UIToolKit.Menu import *
+from UIToolKit.MainMenu import *
 from UIToolKit.ImageUtils import *
 
 pygame.init()
 
 def Main():
     start = time.time()
-    menu = Menu(3, "")
-    menu.MakeButtons()
+    mainMenu = MainMenu()
     #menu
     #ingame
     #paused game
@@ -19,9 +18,9 @@ def Main():
 
     while True:
         for event in pygame.event.get(): # event handling loop
-            menu.handleInputs(event)
+            mainMenu.handleInputs(event)
 
-        menu.draw()
+        mainMenu.draw()
 
         pygame.display.flip()
         time.sleep(0.2)
